+++ db.js
@@ -1,0 +1,12 @@
+const Sequelize = require("sequelize");
+
+const databaseUrl =
+  process.env.DATABASE_URL ||
+  "postgres://postgres:secret@localhost:5432/postgres";
+const db = new Sequelize(databaseUrl);
+
+db.sync({ force: false })
+  .then(() => console.log("Database synced"))
+  .catch(console.error);
+
+module.exports = db;
