+++ user/router.js
@@ -1,0 +1,5 @@
+const { Router } = require("express");
+const User = require("./model");
+const router = new Router();
+
+module.exports = router;
